change variable naming so it correspond to its path not only to its occurence number; allow for custom naming dedupe function; precompile module to es5; update readme
