--- .editorconfig
@@ -1,7 +1,0 @@
-root = true
-
-[*.{js,sql,md}]
-indent_style = space
-indent_size = 2
-max_line_length = 120
-quote_type = single
