--- .eslintrc
@@ -1,13 +1,0 @@
-{
-    "extends": ["airbnb-base"],
-    "rules": {
-      "no-console": "off",
-      "global-require": "off",
-      "no-plusplus": "off",
-      "no-param-reassign": "warn"
-    },
-    "env": {
-      "mocha": true
-    }
-}
-
