+++ .gitignore
@@ -1,3 +1,5 @@
+lib/
+
 #Pycharm charm
 .idea
